codegen: Convert most methods returning non-pointer to out-param convention
This means that they return gboolean indicating whether or not the method
succeeded or failed, take a GError argument and take the out-parameter
as one of their last arguments. We mark this parameter as an out-parameter
so that gobject-introspection knows what to do with it.

This will then allow us to handle exceptions in these methods.