torch-gobject: Add TorchCallbackData and internal helpers
This is a boxed type that can be used to store callback information
in a slightly more efficient way than GClosure. The reason for using
this over GClosure is that we don't really need to worry about custom
marshalling function - its the C caller's responsibility to get it right,
we just need to care about keeping the callback data alive as the boxed type
might be passed around in the bindings.